api/v1/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
rq==1.16.2
redis==5.2.1
python-dotenv==1.0.1
python-multipart==0.0.20
